more work on the bot

+++ strangelove/core.py
@@ -1,16 +1,16 @@
 import asyncio as aio
-import aiohttp
 
-from strangelove.classify import classify, compare
+from strangelove.classify import classify
 from strangelove.db import connect, Movie
 from strangelove import thepiratebay
 
 
 class Core:
 
-    def __init__(self):
+    def __init__(self, session):
         self._db = connect()
-        self._session = aiohttp.ClientSession()
+        print(self._db)
+        self._session = session
 
     async def search(self, m):
         return await thepiratebay.search(m.format_search(),
@@ -18,31 +18,30 @@ class Core:
 
     async def add_movie(self, title, year=None):
         m = Movie(title, year)
-        try:
-            torrents = await self.search(m)
-        except:
-            pass
-        else:
-            kls = classify(torrents)
-            for cat in kls.values():
-                for t in cat:
-                    m.torrents.append(t)
-            self._db.add(m)
-            self._db.commit()
-            return kls
+        torrents = await self.search(m)
+        m.torrents = list(torrents)
+        self._db.add(m)
+        self._db.commit()
+        return len(torrents), classify(torrents)
 
     def rm_movie(self, title):
         m = self._db.query(Movie).filter(Movie.title.ilike(title)).first()
-        self._db.delete(m)
-        self._db.commit()
+        if m is not None:
+            self._db.delete(m)
+            self._db.commit()
+        else:
+            raise ValueError
 
     def list_movies(self):
         return self._db.query(Movie).all()
 
     async def check_movie(self, m):
         torrents = await self.search(m)
-        diff = compare(m.torrents, torrents)
-        return (m, diff)
+        cache = set(m.torrents)
+        diff = torrents - cache
+        if len(diff) > 0:
+            classify(diff)
+            # TODO more
 
     async def check_movies(self):
         movies = self._db.query(Movie).all()
@@ -52,6 +51,3 @@ class Core:
             return [f.result() for f in futs]
         else:
             return []
-
-    def __del__(self):
-        self._session.close()
